+++ src/main/kotlin/com/dp/advancedgunnerycontrol/weaponais/tags/OpportunistTag.kt
@@ -1,6 +1,8 @@
 package com.dp.advancedgunnerycontrol.weaponais.tags
 
 import com.dp.advancedgunnerycontrol.utils.determineIfShotWillHit
+import com.dp.advancedgunnerycontrol.utils.effectiveCollRadius
+import com.dp.advancedgunnerycontrol.weaponais.isAimable
 import com.dp.advancedgunnerycontrol.weaponais.isOpportuneTarget
 import com.fs.starfarer.api.combat.CombatEntityAPI
 import com.fs.starfarer.api.combat.ShipAPI
@@ -23,8 +25,7 @@ class OpportunistTag(weapon: WeaponAPI) : WeaponAITagBase(weapon) {
     }
 
     override fun shouldFire(entity: CombatEntityAPI, predictedLocation: Vector2f): Boolean {
-
-        if(!determineIfShotWillHit(predictedLocation, entity.collisionRadius.times(0.5f), weapon)){
+        if(isAimable(weapon) && !determineIfShotWillHit(predictedLocation, effectiveCollRadius(entity), weapon)){
             return false
         }
         return isOpportuneTarget(entity, predictedLocation, weapon)
